fix(render): BlendMesh viajaba como float y disparaba blend aditivo de mas
En IDA el parametro `a5` de sub_440D50 (y `a4` de sub_441E00) es un ENTERO:
el indice de la malla a blendear, o -1 para "ninguna".  El port lo tipo como
`float` y quedaron conviviendo DOS codificaciones incompatibles en los 89
call sites:

  - bit-pattern:  *(float *)(o + 100)   -> los bits del int llegan intactos
  - numerica:     -1.0f                 -> bits 0xBF800000 = int -1082130432

La rama del decompile es `else if ( a5 <= -2 || mesh->Texture == a5 )", que
enciende EnableAlphaBlend (aditivo) y pinta con BlendMeshLight.  Con `-1.0f`
el test `<= -2` daba TRUE, asi que 9 call sites que en el binario texturan
normal aca se dibujaban en aditivo.

Sintoma reportado: el Grand Soul Shield se veia demasiado blanco.  sub_504130
case 607 dibuja la malla 0 (jh03.tga: marco bronce con panel interior claro)
con `sub_440D50(red, 0, 2, 1.0, -1, ...)`.  Nosotros la sumabamos en aditivo
sobre el resto con color (1,1,1) -> blanco quemado.  Verificado en runtime con
una sonda: las 3 mallas se dibujaban y con sus texturas correctas (533/535/536,
GL 276/278/279), asi que no era carga ni bind.

El parametro pasa a `int` en las tres funciones de la cadena (sub_441E00,
sub_440D50, sub_4414D0) y los 89 call sites se convierten:
  *(float *)(X + 100) -> *(int *)(X + 100)      (45)
  -1.0f -> -1   0.0f -> 0   3.0f -> 3           (20)
  1.4013e-45f -> 1   2.8026e-45f -> 2   4.2039e-45f -> 3   (denormales = bits)

Dentro de sub_440D50 el parametro ademas se REUSABA como tag de modo (se le
asignaban los sentinelas RENDER_MODE_*); eso pasa a una variable propia
`renderMode` con los valores de `v47` del decompile (1/2/4/64).

HiddenMesh (`a8`) tiene la misma dualidad pero es inocua: los dos encodings
dan "ninguna malla oculta" porque los indices reales son >= 0.  Queda anotado.

+++ src/Render/Entity_AnimationLegacy.cpp
@@ -103,7 +103,7 @@ int __cdecl FUN_004faa70(int param_1, char param_2, int param_3) {
         FUN_004404e0(this_, (int)puVar3, (float*)(param_1+0x118),
                      (float*)(param_1+0x124), (float*)(param_1+0x130), param_2);
         FUN_00441e00(this_, 0x40,
-                     *(float*)(param_1+0x168), *(float*)(param_1+100),
+                     *(float*)(param_1+0x168), *(int *)(param_1+100),
                      *(float*)(param_1+0x68),  *(float*)(param_1+0x6c),
                      *(float*)(param_1+0x70),  *(float*)(param_1+0x58), 0xffffffff);
         FUN_004fa930(param_1, (int)this_);
